+++ android/src/androidTest/java/com/magnet/mmx/client/api/MMXChannelSingleSessionTest.java
@@ -19,6 +19,7 @@ import android.util.Log;
 import com.magnet.max.android.User;
 import com.magnet.mmx.client.api.MMXChannel.FailureCode;
 import com.magnet.mmx.client.utils.ChannelHelper;
+import com.magnet.mmx.client.utils.ExecMonitor;
 import com.magnet.mmx.client.utils.MaxHelper;
 import com.magnet.mmx.client.utils.TestCaseTimer;
 import com.magnet.mmx.client.utils.TestConstants;
@@ -29,13 +30,10 @@ import java.util.List;
 import java.util.Set;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
-import org.junit.After;
 import org.junit.AfterClass;
-import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TestName;
 import org.junit.runner.RunWith;
 
 import static junit.framework.Assert.fail;
