+++ android/src/androidTest/java/com/magnet/mmx/client/api/MMXChannelTest.java
@@ -19,18 +19,16 @@ import android.util.Log;
 import com.magnet.max.android.User;
 import com.magnet.mmx.client.api.MMXChannel.FailureCode;
 import com.magnet.mmx.client.utils.ChannelHelper;
+import com.magnet.mmx.client.utils.ExecMonitor;
 import com.magnet.mmx.client.utils.MaxHelper;
 import com.magnet.mmx.client.utils.TestCaseTimer;
 import com.magnet.mmx.client.utils.TestConstants;
 import com.magnet.mmx.client.utils.UserHelper;
 import java.util.List;
-import org.junit.After;
 import org.junit.AfterClass;
-import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TestName;
 import org.junit.runner.RunWith;
 
 import static junit.framework.Assert.assertEquals;
