+++ android/src/androidTest/java/com/magnet/mmx/client/api/MMXMessageSingleSessionTest.java
@@ -20,6 +20,7 @@ import com.magnet.max.android.MaxCore;
 import com.magnet.max.android.User;
 import com.magnet.mmx.client.api.MMXMessage.InvalidRecipientException;
 import com.magnet.mmx.client.common.Log;
+import com.magnet.mmx.client.utils.ExecMonitor;
 import com.magnet.mmx.client.utils.FailureDescription;
 import com.magnet.mmx.client.utils.MaxHelper;
 import com.magnet.mmx.client.utils.MessageHelper;
@@ -32,16 +33,10 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicLong;
-import java.util.concurrent.atomic.AtomicReference;
-import org.junit.After;
 import org.junit.AfterClass;
-import org.junit.Before;
 import org.junit.BeforeClass;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TestName;
 import org.junit.runner.RunWith;
 
 import static junit.framework.Assert.assertEquals;
