+++ android/src/androidTest/java/com/magnet/mmx/client/api/MMXMessageTest.java
@@ -17,19 +17,17 @@ package com.magnet.mmx.client.api;
 import android.support.test.runner.AndroidJUnit4;
 import com.magnet.max.android.User;
 import com.magnet.mmx.client.common.Log;
+import com.magnet.mmx.client.utils.ExecMonitor;
 import com.magnet.mmx.client.utils.MaxHelper;
 import com.magnet.mmx.client.utils.TestCaseTimer;
 import com.magnet.mmx.client.utils.TestConstants;
 import com.magnet.mmx.client.utils.UserHelper;
 import java.util.HashMap;
 import java.util.HashSet;
-import org.junit.After;
 import org.junit.AfterClass;
-import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TestName;
 import org.junit.runner.RunWith;
 
 import static junit.framework.Assert.assertEquals;
